Move LEQuad to dedicated folder
