+++ drones/lequad.hpp
