+++ rules_common.mk
@@ -87,6 +87,8 @@ LIB_SRCS += drivers/state_display.cpp
 LIB_SRCS += drivers/state_display_megafly_rev4.cpp
 LIB_SRCS += drivers/state_display_sparky_v2.cpp
 
+LIB_SRCS += drones/lequad.cpp
+
 LIB_SRCS += hal/common/dbg.cpp
 LIB_SRCS += hal/common/led_gpio.cpp
 LIB_SRCS += hal/common/file.cpp
@@ -153,9 +155,3 @@ LIB_SRCS += util/print_util.cpp
 LIB_SRCS += util/quick_trig.cpp
 LIB_SRCS += util/raytracing.cpp
 LIB_SRCS += util/string_util.cpp
-
-
-####################################################################################################
-# SAMPLE PROJECT SOURCE FILES
-####################################################################################################
-LIB_SRCS += sample_projects/LEQuad/lequad.cpp
