+++ sample_projects/LEQuad/main_sparky_v2.cpp
@@ -57,7 +57,7 @@
 #include "hal/dummy/pwm_dummy.hpp"
 #include "hal/stm32/spi_stm32.hpp"
 
-#include "sample_projects/LEQuad/lequad.hpp"
+#include "drones/lequad.hpp"
 
 #include "simulation/dynamic_model_quad_diag.hpp"
 #include "simulation/simulation.hpp"
