+++ sample_projects/LEQuad/proj_avr32/Makefile
@@ -15,7 +15,6 @@ MAVRIC_SRC=../
 
 # Sources files for project folder (*.c and *.cpp)
 SRCS += sample_projects/LEQuad/main_avr32.cpp
-SRCS += sample_projects/LEQuad/lequad.cpp
 
 # ------------------------------------------------------------------------------
 # MAVRIC LIBRARY
