+++ sample_projects/LEQuad/proj_linux/Makefile
@@ -12,7 +12,6 @@ PROJ_NAME=LEQuadLI
 # ------------------------------------------------------------------------------
 # Project source files (*.c and *.cpp)
 LIB_SRCS += sample_projects/LEQuad/main_linux.cpp
-LIB_SRCS += sample_projects/LEQuad/lequad.cpp
 
 # ------------------------------------------------------------------------------
 # MAVRIC LIBRARY
