+++ sample_projects/LEQuad/proj_stm32/Makefile
@@ -13,7 +13,6 @@ PROJ_NAME=LEQuadST
 # Project source files (*.c and *.cpp)
 # LIB_SRCS += sample_projects/LEQuad/main_stm32.cpp
 LIB_SRCS += sample_projects/LEQuad/main_sparky_v2.cpp
-LIB_SRCS += sample_projects/LEQuad/lequad.cpp
 
 # ------------------------------------------------------------------------------
 # MAVRIC LIBRARY
