+++ sample_projects/LEQuad/untested/proj_sparky_chibios/Makefile
@@ -133,8 +133,7 @@ CSRC += $(STARTUPSRC) \
 # setting.
 CPPSRC += $(BOARDCPPSRC) \
 		  $(addprefix ${MAVRIC_LIB}/, $(LIB_SRCS)) \
-         ../../main_chibios_sparky.cpp \
-		 ../../lequad.cpp
+         ../../main_chibios_sparky.cpp
 
 # C sources to be compiled in ARM mode regardless of the global setting.
 # NOTE: Mixing ARM and THUMB mode enables the -mthumb-interwork compiler
